% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_standard.R
\name{bootstrap_standard}
\alias{bootstrap_standard}
\title{Title}
\usage{
bootstrap_standard(
  olsobj,
  irfobj,
  h,
  nrep,
  method = "standard",
  standard_factor,
  efron_percentile_quantiles,
  hall_percentile_quantiles,
  nrep_inside_boot,
  bootstrap_seed,
  display_progress_bar = TRUE
)
}
\arguments{
\item{method}{String for method of computing standard errors.}

\item{display_progress_bar}{}
}
\description{
Title
}
