% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_wild.R
\name{bootstrap_wild}
\alias{bootstrap_wild}
\title{Compute confidence intervals of a structural VAR(p) model using the wild bootstrap}
\usage{
bootstrap_wild(
  olsobj,
  irfobj,
  h,
  nrep,
  method = "standard",
  standard_factor,
  efron_percentile_quantiles,
  hall_percentile_quantiles,
  nrep_inside_boot,
  bootstrap_seed,
  display_progress_bar = TRUE
)
}
\arguments{
\item{olsobj}{Least-squares object from running \code{olsvarc()}}

\item{irfobj}{IRF object from running ``.}

\item{method}{String for method of computing standard errors.}

\item{display_progress_bar}{}
}
\description{
Compute confidence intervals of a structural VAR(p) model using the wild bootstrap
}
