% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_fcast_error_decomp.R
\name{compute_fcast_error_decomp}
\alias{compute_fcast_error_decomp}
\title{Compute forecast error variance decompositions}
\usage{
compute_fcast_error_decomp(solvar, k, p, h, varpos)
}
\arguments{
\item{solvar}{A list that is output from olsvarc.R.}

\item{k}{A number that represents the number of series: y_t is k by 1.}

\item{p}{A number that represents the order of the VAR process; that is, the number of lags.}

\item{h}{A number that represents the horizon.}

\item{varpos}{A number that represents position of the series.}
}
\value{
A number that represents the percent
}
\description{
Compute forecast error variance decompositions
}
\examples{
myf1 <- fun_compute_fcast_error_decomp(solvar = solvar, k = k, p = p, h = 1, varpos = 4)
}
