% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_fcast_error_var_decomp.R
\name{compute_fcast_error_var_decomp}
\alias{compute_fcast_error_var_decomp}
\title{Compute how much of the prediction mean squared error is accounted for by each structural shock}
\usage{
compute_fcast_error_var_decomp(solvar, h, var_name, eps = 1e-04)
}
\arguments{
\item{solvar}{A list that is output from running \code{olsvarc()}.}

\item{h}{A number that represents the horizon.}

\item{eps}{A number to check the absolute difference between percent of h-step ahead forecast error variance explained.}

\item{varpos}{A string that names the variable to be explained.}
}
\value{
If h is finite, a vector of numbers that represents the percent explained; otherwise, a tibble of the percent explained at different horizons.
}
\description{
The forecast error variance decomposition of the structural model computes
how much of prediction mean squared error of y_{t+h} at horizon h=1,...,H is
accounted for by each structural shock.
}
\examples{
y_stock_market <- kilianLutkepohlCh04Table4_1 |> select(all_of(var_order_stock_market))
sol_stock_market <- olsvarc(y_stock_market, p = 24)
compute_fcast_error_var_decomp(solvar = sol_stock_market, h = 3, var_name = "dd")
}
