% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_hist_decomp.R
\name{compute_hist_decomp}
\alias{compute_hist_decomp}
\title{Compute historical decomposition}
\usage{
compute_hist_decomp(olsvarc_obj, date, var_order)
}
\arguments{
\item{olsvarc_obj}{Object created by \code{olsvarc()}}

\item{date}{The \code{date} from the original data (trimmed to T - p)}

\item{var_order}{Order of variables for structural VAR model}
}
\description{
Compute historical decomposition
}
