% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detrendcl.R
\name{detrendcl}
\alias{detrendcl}
\title{Remove the mean (constant) or linear trend (linear) from a vector.}
\usage{
detrendcl(x, tt)
}
\arguments{
\item{x}{A vector considered a time series.}

\item{tt}{A string that indicates trend type, "constant" or "linear".}
}
\value{
A vector that has removed the mean or linear trend from \code{x}. That is,
for \code{y <- detrend(x, tt = "linear")}, then \code{x} equals \code{y} + linear trend or
linear trend = \code{x - y}.
}
\description{
Missing values are ignored. For example, suppose the change in production is
the variable of interest. Generating the variable generates a missing value
in the first position. A linear trend using \code{detrendcl()} can still be used.
But missing values in the middle of the series will also be ignored, which is
probably not the desired outcome. A message is thrown when a missing value is
encountered.
}
\details{
See the pracma package for more robust commands.
}
\examples{
x <- 1:11
x - detrend(x, tt = "constant")
detrend(x, tt = "linear")
}
