% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irf2.R
\name{irf2}
\alias{irf2}
\title{Compute second-stage impulse response functions}
\usage{
irf2(
  y,
  x,
  p,
  block_length,
  standard_deviation_factor,
  nrep = 2000,
  boot_seed = 676,
  cumeffect = FALSE
)
}
\arguments{
\item{y}{Vector or matrix of time series}

\item{x}{Vector or matrix of shock series}

\item{p}{A positive integer that represents the horizon}

\item{block_length}{Block length of d}

\item{standard_deviation_factor}{Non-negative number to scale up the standard deviation of the bootstrap draws.}

\item{nrep}{Number that represents the number of bootstrap draws}

\item{boot_seed}{Number to seed the bootstrap}

\item{cumeffect}{Logical to indicate whether to cummulate the impulse response}
}
\value{
A tibble with second-stage impulse response and upper- and lower-ends that define the confidence interval.
}
\description{
Compute second-stage impulse response functions
}
\examples{
library(dplyr)
library(tibble)
library(ggplot2)
library(lubridate)
y <- kilianLutkepohlCh12Figure12_5 |>
  dplyr::rename(oilsupply = oil_supply,
                aggdemand = agg_demand) |>
  dplyr::select(-date)
var_order <- c("oilsupply", "aggdemand", "rpoil")
negative_shocks <- "oilsupply"
var_cumsum <- c("response_shock_oilsupply_oilsupply",
                "response_shock_oilsupply_aggdemand",
                "response_shock_oilsupply_rpoil")
p <- 24
sol <- olsvarc(y, p = p)
# Estimate structural shocks
Ehat <- solve(t(chol(sol$SIGMAhat))) \%*\% sol$Uhat
# Create tibble of structural shocks
dates <- kilianLutkepohlCh12Figure12_5|>
  slice(seq(from = p + 1, to = nrow(kilianLutkepohlCh12Figure12_5), by = 1)) |>
  pull(date)
dat_structural_shocks <- as_tibble(t(Ehat)) |>
  rename_with(.fn = function(x) paste0("shock_", x), .cols = everything()) |>
  mutate(date = dates)

# Generate some random data to work with
dat_structural_shocks <- dat_structural_shocks |>
  mutate(somey = rnorm(n(), mean = 3, sd = 1))

# Create 2nd-stage IRFs
dat_irf2 <- irf2(y = dat_structural_shocks$somey, x = dat_structural_shocks$shock_oilsupply,
                 p = 12, block_length = 12, standard_deviation_factor = 2, boot_seed = 1, nrep = 500)

ggplot(data = dat_irf2) +
  geom_line(mapping = aes(x = horizon, y = irf2)) +
  geom_line(mapping = aes(x = horizon, y = irf2_lo), color = "blue", linetype = "dotted") +
  geom_line(mapping = aes(x = horizon, y = irf2_hi), color = "blue", linetype = "dotted")
}
