% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irfvar.R
\name{irfvar}
\alias{irfvar}
\title{Compute structural impulse responses.}
\usage{
irfvar(Ahat, B0inv, p, h, var_order, var_cumsum = NULL, negative_shocks = NULL)
}
\arguments{
\item{Ahat}{A matrix of VAR parameters.}

\item{B0inv}{A matrix that corresponds to the structural impact multiplier matrix.}

\item{p}{A number that represents the number of lags.}

\item{h}{A number that represents the horizon.}

\item{var_order}{A string of variable names that specify the VAR ordering.}

\item{var_cumsum}{A character vector indicating shock_response series to cumulate.}

\item{negative_shocks}{A character vector indicating negative shocks.}

\item{return_tidy}{A logical indicator for whether to return a tidy data set (TRUE) or a matrix (FALSE).}
}
\value{
A matrix of impulse responses
}
\description{
Compute structural impulse responses.
}
\examples{
IRF <- irfvar(sol$Ahat, B0inv, p, h)
}
