% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kilian2009AER}
\alias{kilian2009AER}
\title{Data used in Lutz Kilian's "Not All Oil Price Shocks Are Alike: Disentangling Demand and Supply Shocks in the Crude Oil Market," American Economic Review, 99 (3): 1053--69.}
\format{
\subsection{\code{kilian2009AER}}{

A data frame with 419 rows (data from 1973-02-01 to 2007-12-01) and 4 columns:
\describe{
\item{d_production}{Change in oil production}
\item{real_economic_activity}{Real economic activity}
\item{real_price_oil}{Real price of oil}
\item{date}{Date object (constructed with \code{lubridate::ymd()})}
}
}
}
\source{
\url{https://www.aeaweb.org/articles?id=10.1257/aer.99.3.1053}
}
\usage{
kilian2009AER
}
\description{
Data used in Lutz Kilian's "Not All Oil Price Shocks Are Alike: Disentangling Demand and Supply Shocks in the Crude Oil Market," American Economic Review, 99 (3): 1053--69.
}
\keyword{datasets}
