% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kilianLutkepohlCh02Macro}
\alias{kilianLutkepohlCh02Macro}
\title{Data used in section 2.3, figure 2.1 of Kilian and Lutkepohl's book "Structural Vector Autoregressive Analysis."}
\format{
\subsection{\code{kilianLutkepohlCh02Macro}}{

A data frame with 213 rows (1954 quarter 4 to 2007 quarter 4) and 9 columns:
\describe{
\item{year}{Year (integer)}
\item{quarter}{Quarter (integer)}
\item{qtr}{Quarter (Date object constructed with \code{lubridate::yq()})}
\item{date}{Date object}
\item{gnp_deflator}{GNP deflator}
\item{infl}{Inflation measured by 100 times the difference in the log of the GNP deflator}
\item{rgnp}{Real economic activity}
\item{drgnp}{Difference in log of real GNP times 100}
\item{irate}{Quarterly average of monthly values of the federal funds rate}
}
}
}
\source{
Lutz Kilian's website: \url{https://sites.google.com/site/lkilian2019/textbook/code}
}
\usage{
kilianLutkepohlCh02Macro
}
\description{
The data are US macro data on real GNP, the federal funds rate, and inflation.
The data are quarterly.
}
\keyword{datasets}
