% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kilianLutkepohlCh12Figure12_5}
\alias{kilianLutkepohlCh12Figure12_5}
\title{Data used in section 12.13, figure 12.5 of Kilian and Lutkepohl's book "Structural Vector Autoregressive Analysis."}
\format{
\subsection{\code{kilianLutkepohlCh12Figure12_5}}{

A data frame with 419 rows (data from 1973-02-01 to 2007-12-01) and 4 columns:
\describe{
\item{date}{Date object, monthly (constructed with \code{lubridate::ymd()})}
\item{oil_supply}{Percent change in world crude-oil production}
\item{agg_demand}{Index of real economic activity}
\item{rpoil}{Real price of oil}
}
}

\subsection{\code{kilianLutkepohlCh09Figure9_1}}{

A data frame with 162 rows (data from 1973q1 to 2013q2) and 6 columns:
\describe{
\item{year}{Year}
\item{quarter}{Quarter}
\item{date}{Date object, monthly (constructed with \code{lubridate::ymd()})}
\item{drpoil}{Percent change in the real WTI price of crude oil}
\item{infl}{Inflation rate measured by the US GDP deflator}
\item{drgdp}{Growth rate of US real GDP}
}
}
}
\source{
Lutz Kilian's website: \url{https://sites.google.com/site/lkilian2019/textbook/code}

Lutz Kilian's website: \url{https://sites.google.com/site/lkilian2019/textbook/code}
}
\usage{
kilianLutkepohlCh12Figure12_5

kilianLutkepohlCh12Figure12_5
}
\description{
The data are monthly data on world crude-oil production, global real economic activity, and the real price of oil.

The data are monthly data on world crude-oil production, global real economic activity, and the real price of oil.
}
\keyword{datasets}
