% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonally_adjust_monthly.R
\name{seasonally_adjust_monthly}
\alias{seasonally_adjust_monthly}
\title{Seasonally adjust monthly data}
\usage{
seasonally_adjust_monthly(x, date)
}
\arguments{
\item{x}{A vector.}

\item{date}{A vector of monthly dates.}

\item{...}{Other options passed to seasonal::seas().}
}
\value{
A vector.
}
\description{
\code{seasonally_adjust_monthly()} uses the function \code{seas} from the \code{seasonal} package,
which "calls the automatic procedures of X-13ARIMA-SEATS to perform a seasonal adjustment that works well in most circumstances."
}
\examples{
seasonally_adjust(data$some_series, data$date)
}
