% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_wild.R
\name{bootstrap_wild}
\alias{bootstrap_wild}
\title{Compute confidence intervals of a structural VAR(p) model using the wild bootstrap}
\usage{
bootstrap_wild(
  olsobj,
  irfobj,
  h,
  nrep,
  method = "standard",
  standard_factor,
  efron_percentile_quantiles,
  hall_percentile_quantiles,
  nrep_inside_boot,
  bootstrap_seed,
  display_progress_bar = TRUE
)
}
\arguments{
\item{olsobj}{Least-squares object from running \code{olsvarc()}}

\item{irfobj}{IRF object from running \code{irfvar()}.}

\item{h}{Horizon for IRFs}

\item{nrep}{Positive integer for number of bootstrap replications}

\item{method}{String for method of computing standard errors.}

\item{bootstrap_seed}{Positive integer for seed}

\item{display_progress_bar}{Logical for whether to display the progress bar}
}
\value{
A tibble dataframe containing upper and lower ends of the confidence intervals associated with impulse response functions.
}
\description{
Compute confidence intervals of a structural VAR(p) model using the wild bootstrap
}
\examples{
y <- kilianLutkepohlCh12Figure12_5 |>
  dplyr::rename(oilsupply = oil_supply,
                aggdemand = agg_demand) |>
  dplyr::select(-date)
var_order <- c("oilsupply", "aggdemand", "rpoil")
negative_shocks <- "oilsupply"
var_cumsum <- c("response_shock_oilsupply_oilsupply",
                "response_shock_oilsupply_aggdemand",
                "response_shock_oilsupply_rpoil")
sol <- olsvarc(y, p = 24)
irf <- irfvar(Ahat = sol$Ahat, B0inv = t(chol(sol$SIGMAhat)), p = sol$p, h = 15,
              var_order = var_order,
              negative_shocks = negative_shocks,
              var_cumsum = var_cumsum)
dat_ci <- bootstrap_wild(sol, irf, nrep = 500, method = "standard", standard_factor = 2.0)
plot(irf$irf_tidy$horizon, irf$irf_tidy$response_shock_oilsupply_oilsupply, type = "l", ylim = c(-2, 0))
lines(dat_ci$horizon, dat_ci$response_shock_oilsupply_oilsupply_lo, lty = "dotted", col = "blue")
lines(dat_ci$horizon, dat_ci$response_shock_oilsupply_oilsupply_hi, lty = "dotted", col = "blue")

# Another example ---------------------------------------------------------

y <- kilianLutkepohlCh02Macro |>
  dplyr::select(drgnp, infl)

sol <- olsvarc(y, p = 4)
irf <- irfvar(Ahat = sol$Ahat, B0inv = t(chol(sol$SIGMAhat)), p = sol$p, h = 8,
              var_order = c("drgnp", "infl"))
dat_ci <- bootstrap_standard(sol, irf, nrep = 500, method = "standard", standard_factor = 2.0)
plot(irf$irf_tidy$horizon, irf$irf_tidy$response_shock_infl_drgnp, type = "l", ylim = c(-0.15, 0.15))
lines(dat_ci$horizon, dat_ci$response_shock_infl_drgnp_lo, lty = "dotted", col = "blue")
lines(dat_ci$horizon, dat_ci$response_shock_infl_drgnp_hi, lty = "dotted", col = "blue")
}
