% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_fcast_error_var_decomp.R
\name{compute_fcast_error_var_decomp}
\alias{compute_fcast_error_var_decomp}
\title{Compute how much of the prediction mean squared error is accounted for by each structural shock}
\usage{
compute_fcast_error_var_decomp(solvar, k, p, h, varpos)
}
\arguments{
\item{solvar}{A list that is output from running \code{olsvarc()}.}

\item{k}{A number that represents the number of series: y_t is k by 1.}

\item{p}{A number that represents the order of the VAR process; that is, the number of lags.}

\item{h}{A number that represents the horizon.}

\item{varpos}{A number that represents position of the series.}
}
\value{
A number that represents the percent
}
\description{
The forecast error variance decomposition of the structural model computes
how much of prediction mean squared error of y_{t+h} at horizon h=1,...,H is
accounted for by each structural shock.
}
\examples{
myf1 <- compute_fcast_var_error_var_decomp(solvar = solvar, k = k, p = p, h = 1, varpos = 4)
}
