% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_hist_decomp.R
\name{compute_hist_decomp}
\alias{compute_hist_decomp}
\title{Compute historical decomposition}
\usage{
compute_hist_decomp(olsvarc_obj, rownames, var_order)
}
\arguments{
\item{olsvarc_obj}{Object created by \code{olsvarc()}}

\item{rownames}{Names for returned T - p tibble (most likely \code{date} from the original data)}

\item{var_order}{Order of variables for structural VAR model}
}
\description{
Compute historical decomposition
}
