% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detrendcl.R
\name{detrendcl}
\alias{detrendcl}
\title{Remove the mean (constant) or linear trend (linear) from a vector.}
\usage{
detrendcl(x, tt)
}
\arguments{
\item{x}{A vector considered a time series.}

\item{tt}{A string that indicates trend type, "constant" or "linear".}
}
\value{
A vector that has removed the mean or linear trend from \code{x}. That is,
for \code{y <- detrend(x, tt = "linear")}, then \code{x} equals \code{y} + linear trend or
linear trend = \code{x - y}.
}
\description{
See the pracma package for more robust commands.
}
\examples{
x <- 1:11
x - detrend(x, tt = "constant")
detrend(x, tt = "linear")
}
