% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irfvar.R
\name{irfvar}
\alias{irfvar}
\title{Compute structural impulse responses.}
\usage{
irfvar(Ahat, B0inv, p, h)
}
\arguments{
\item{Ahat}{A matrix of VAR parameters.}

\item{B0inv}{A matrix that corresponds to the structural impact multiplier matrix.}

\item{p}{A number that represents the number of lags.}

\item{h}{A number that represents the horizon.}
}
\value{
A matrix of impulse responses
}
\description{
Compute structural impulse responses.
}
\examples{
IRF <- irfvar(sol$Ahat, B0inv, p, h)
}
