% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olsvarc.R
\name{olsvarc}
\alias{olsvarc}
\title{Estimate a VAR model with an intercept using OLS.}
\usage{
olsvarc(y, p)
}
\arguments{
\item{y}{A matrix of data.}

\item{p}{A positive integer, indicating the number of lags.}
}
\value{
A list with components\tabular{ll}{
\code{Vhat} \tab A numeric vector of parameter estimates \cr
\tab \cr
\code{Ahat} \tab OLS estimates of A_i, i = 1,..., p, where each A_i is K by K \cr
\tab \cr
\code{SIGMAhat} \tab Consistent estimator of the innovation covariance matrix \cr
\tab \cr
\code{Uhat} \tab Estimated least-squares residuals \cr
\tab \cr
\code{Z} \tab Constructed data matrix \cr
\code{Y} \tab Constructed data matrix \cr
\code{K} \tab Number of series in y \cr
\code{y} \tab The original dataset used as an argument \cr
}
}
\description{
Details are provided on page 31 of Kilian, Lutz, and Helmut Lütkepohl. \emph{Structural Vector Autoregressive Analysis.}
}
\examples{
sol <- olsvarc(y, p)
}
